# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"Joe Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/reservations/{idReservation}":{"get":{"tags":["reservation-controller"],"operationId":"getReservationById","parameters":[{"name":"idReservation","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}},"put":{"tags":["reservation-controller"],"operationId":"updateReservation","parameters":[{"name":"idReservation","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}}},"/reservations":{"post":{"tags":["reservation-controller"],"operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}}},"/flights":{"post":{"tags":["flight-controller"],"operationId":"createFlight","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Flight"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Flight"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/flights/{idFlight}":{"get":{"tags":["flight-controller"],"operationId":"getFlightById","parameters":[{"name":"idFlight","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Flight"}}}}}}}},"components":{"schemas":{"Passenger":{"type":"object","properties":{"name":{"type":"string"},"seat":{"type":"string"}}},"Reservation":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string"},"flight":{"type":"string"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/Passenger"}}}},"Airport":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string"},"location":{"type":"string"}}},"Flight":{"type":"object","properties":{"id":{"type":"string"},"aircraft":{"type":"string"},"departure":{"type":"string","format":"date-time"},"airport":{"$ref":"#/components/schemas/Airport"},"seat_map":{"type":"array","items":{"$ref":"#/components/schemas/Seat"}}}},"Seat":{"type":"object","properties":{"seat":{"type":"string"},"status":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}